+++ app/src/main/java/com/alpha/myeyecare/presentation/ui/detailScreen/SetupReminderViewModel.kt
@@ -3,7 +3,6 @@ package com.alpha.myeyecare.presentation.ui.detailScreen
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.alpha.myeyecare.domain.model.ReminderDetails
-import com.alpha.myeyecare.domain.useCases.CheckReminderStatusUseCase
 import com.alpha.myeyecare.domain.useCases.GetReminderDetailsUserCase
 import com.alpha.myeyecare.domain.useCases.SaveReminderUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -15,7 +14,6 @@ import javax.inject.Inject
 
 @HiltViewModel
 class SetupReminderViewModel @Inject constructor(
-    private val checkReminderStatusUseCase: CheckReminderStatusUseCase,
     private val getReminderDetailsUserCase: GetReminderDetailsUserCase,
     private val saveReminderUseCase: SaveReminderUseCase,
 ) : ViewModel() {