+++ app/src/main/java/com/alpha/myeyecare/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.alpha.myeyecare.ui.theme
+package com.alpha.myeyecare.presentation.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle