+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased] - YYYY-MM-DD
 ### Added
+- Refactor & Removed unwanted code.
 - Modified code with Clean Architecture.
 - Added the Notification permission check on Home screen & before saving the reminder.
 - Introduced the ViewModel.
 ### Changed
 ### Fixed 
+- Fixed the issue of not clicking Submit button on dialog dismiss.  
 
 ## [release-v1.0] - 19th August, 25 Tuesday
 ### Added